+++ src/manager.cpp
@@ -132,7 +132,7 @@ bool Manager::load_network() {
         return false;
     }
 
-    int secs = 0;
+    double secs = 0;
     if (m_conf.Get("keep_alive", secs)) {
         m_net->set_keep_alive(secs);
     }
